refactor: separando função de login e visualizar personagem em dois arquivos diferentes
