+++ game/src/operadores/login.py
@@ -1,9 +1,6 @@
-from database import criar_cursor
+def login(nome,senha,cursor):
 
-def verificar_usuario(nome,senha):
-    conn,cursor = criar_cursor()
-
-    if conn and cursor:
+    if cursor:
         try:
             sql = """select id_usuario 
             from public.usuario 
@@ -14,8 +11,7 @@ def verificar_usuario(nome,senha):
             
             if resultado:
                 id = resultado["id_usuario"]
-                print("Resultado encontrado!")
-                visualizar_personagens(id,cursor)
+                print("Usuario encontrado!")
 
             else:
                 print('Usuario não encontrado')
@@ -25,29 +21,4 @@ def verificar_usuario(nome,senha):
             print(f"Erro ao verificar usuario: {e}")
             return None
 
-        finally:
-            conn.close()
-            cursor.close()
-
-def visualizar_personagens(id,cursor):
-    try:
-        sql = """select distinct nome, raça, nivel
-        from personagem
-        where id = %s """
-
-        cursor.execute(sql,(id))
-        resultado = cursor.fetchall()
-
-        if resultado:
-            print("Personagem(s) encontrado(s): ")
-            for personagem in resultado:
-                print(personagem)
-            
-        else:
-            print("Nenhum personagem encontrado.")
-
-    except Exception as e:
-        print(f"Erro ao encontrar personagens: {e}")
-
-
-verificar_usuario("Alo","123")
+    return id
